+++ alicloud/service_alicloud_vpc.go
@@ -5,10 +5,11 @@ import (
 
 	"fmt"
 
+	"strings"
+
 	"github.com/aliyun/alibaba-cloud-sdk-go/sdk/requests"
 	"github.com/aliyun/alibaba-cloud-sdk-go/services/vpc"
 	"github.com/terraform-providers/terraform-provider-alicloud/alicloud/connectivity"
-	"strings"
 )
 
 type VpcService struct {